requirements.txt
ddddocr==1.4.7
pycryptodome==3.18.0
requests==2.28.2
